+++ node/internal.ts
@@ -20,6 +20,15 @@ import crypto = require('crypto');
 export var _knownVariableMap: { [key: string]: _KnownVariableInfo; } = {};
 
 export var _vault: vm.Vault;
+var _commandCorrelationId: string;
+
+//-----------------------------------------------------
+// Enums
+//-----------------------------------------------------
+export enum IssueSource {
+    CustomerScript = 'CustomerScript',
+    TaskInternal = 'TaskInternal'
+}
 
 //-----------------------------------------------------
 // Validation Checks
@@ -27,7 +36,7 @@ export var _vault: vm.Vault;
 
 // async await needs generators in node 4.x+
 if (semver.lt(process.versions.node, '4.2.0')) {
-    _warning('Tasks require a new agent.  Upgrade your agent or node to 4.2.0 or later');
+    _warning('Tasks require a new agent.  Upgrade your agent or node to 4.2.0 or later', IssueSource.TaskInternal);
 }
 
 //-----------------------------------------------------
@@ -40,6 +49,20 @@ export function _startsWith(str: string, start: string): boolean {
 
 export function _endsWith(str: string, end: string): boolean {
     return str.slice(-end.length) == end;
+}
+
+export function _truncateBeforeSensitiveKeyword(str: string, sensitiveKeywordsPattern: RegExp): string {
+    if(!str) {
+        return str;
+    }
+
+    const index = str.search(sensitiveKeywordsPattern); 
+
+    if (index <= 0) {
+        return str;
+    }
+
+    return `${str.substring(0, index)}...`;
 }
 
 //-----------------------------------------------------
@@ -131,7 +154,7 @@ function _loadLocStrings(resourceFile: string, culture: string): { [key: string]
         }
     }
     else {
-        _warning('LIB_ResourceFile does not exist');
+        _warning('LIB_ResourceFile does not exist', IssueSource.TaskInternal);
     }
 
     return locStrings;
@@ -169,7 +192,7 @@ export function _setResourcePath(path: string, ignoreWarnings: boolean = false):
         if (ignoreWarnings) {
             _debug(_loc('LIB_ResourceFileAlreadySet', path));
         } else {
-            _warning(_loc('LIB_ResourceFileAlreadySet', path));
+            _warning(_loc('LIB_ResourceFileAlreadySet', path), IssueSource.TaskInternal);
         }
     }
 }
@@ -200,7 +223,7 @@ export function _loc(key: string, ...param: any[]): string {
     }
     else {
         if (Object.keys(_resourceFiles).length <= 0) {
-            _warning(`Resource file haven't been set, can't find loc string for key: ${key}`);
+            _warning(`Resource file haven't been set, can't find loc string for key: ${key}`, IssueSource.TaskInternal);
         }
         else {
             _warning(`Can't find loc string for key: ${key}`);
@@ -282,12 +305,12 @@ export function _command(command: string, properties: any, message: string) {
     _writeLine(taskCmd.toString());
 }
 
-export function _warning(message: string): void {
-    _command('task.issue', { 'type': 'warning' }, message);
+export function _warning(message: string, source: IssueSource = IssueSource.TaskInternal): void {
+    _command('task.issue', { 'type': 'warning', 'source': source, 'correlationId': _commandCorrelationId }, message);
 }
 
-export function _error(message: string): void {
-    _command('task.issue', { 'type': 'error' }, message);
+export function _error(message: string, source: IssueSource = IssueSource.TaskInternal): void {
+    _command('task.issue', { 'type': 'error', 'source': source, 'correlationId': _commandCorrelationId }, message);
 }
 
 export function _debug(message: string): void {
@@ -732,6 +755,10 @@ export function _loadData(): void {
         }
     }
     _debug('loaded ' + loaded);
+
+    let correlationId = process.env["COMMAND_CORRELATION_ID"];
+    delete process.env["COMMAND_CORRELATION_ID"];
+    _commandCorrelationId = correlationId ? String(correlationId) : "";
 
     // store public variable metadata
     let names: string[];